permision_final

+++ app/Http/Controllers/BrandController.php
@@ -31,7 +31,7 @@ class BrandController extends Controller
             ->select('list_permisions.ten_hanh_dong')
             ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
             ->toArray(); // Chuyển đổi sang mảng
-        if (in_array('Quản Lý Xe', $check)) {
+        if (in_array('Quản Lý Thương Hiệu', $check)) {
             $data =  $request->all();
             Brand::create($data);
             return response()->json([
@@ -48,21 +48,36 @@ class BrandController extends Controller
     }
     public function del(Request $request)
     {
-        $brand = Brand::where('id', $request->id)->first();
+        $who = Auth::guard('admin')->user();
 
-        if ($brand) {
-            $brand->delete();
+        $check = PermisionDetail::join('list_permisions', 'list_permisions.id', 'permision_details.id_hanh_dong')
+        ->where('id_quyen', $who->id_phan_quyen)
+            ->select('list_permisions.ten_hanh_dong')
+            ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
+            ->toArray(); // Chuyển đổi sang mảng
+        if (in_array('Quản Lý Thương Hiệu', $check)) {
+            $brand = Brand::where('id', $request->id)->first();
 
-            return response()->json([
-                'status'    => true,
-                'message'   => 'Đã xóa Brand thành công !',
-            ]);
+            if ($brand) {
+                $brand->delete();
+
+                return response()->json([
+                    'status'    => true,
+                    'message'   => 'Đã xóa Brand thành công !',
+                ]);
+            } else {
+                return response()->json([
+                    'status'    => false,
+                    'message'   => 'Brand không tồn tại !',
+                ]);
+            }
         } else {
             return response()->json([
-                'status'    => false,
-                'message'   => 'Brand không tồn tại !',
+                'status'    => 0,
+                'message'   => 'Bạn không đủ thẩm quyền để thực hiện chức năng này',
             ]);
         }
+
     }
     public function search(Request $request)
     {
@@ -74,44 +89,88 @@ class BrandController extends Controller
     }
     public function edit(Request $request)
     {
-        $brand = Brand::where('id', $request->id)->first();
-        if ($brand) {
-            return response()->json([
-                'status'    => true,
-                'data'   => $brand,
-            ]);
+        $who = Auth::guard('admin')->user();
+
+        $check = PermisionDetail::join('list_permisions', 'list_permisions.id', 'permision_details.id_hanh_dong')
+        ->where('id_quyen', $who->id_phan_quyen)
+            ->select('list_permisions.ten_hanh_dong')
+            ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
+            ->toArray(); // Chuyển đổi sang mảng
+        if (in_array('Quản Lý Thương Hiệu', $check)) {
+            $brand = Brand::where('id', $request->id)->first();
+            if ($brand) {
+                return response()->json([
+                    'status'    => true,
+                    'data'   => $brand,
+                ]);
+            } else {
+                return response()->json([
+                    'status'    => false,
+                    'message'   => 'Dữ liệu không chính xác !',
+                ]);
+            }
         } else {
             return response()->json([
-                'status'    => false,
-                'message'   => 'Dữ liệu không chính xác !',
+                'status'    => 0,
+                'message'   => 'Bạn không đủ thẩm quyền để thực hiện chức năng này',
             ]);
         }
     }
     public function update(UpdateBrandRequest $request)
     {
-        $brand = Brand::where('id', $request->id)->first();
-        $brand->update([
-            'ten_thuong_hieu'    =>  $request->ten_thuong_hieu,
-            'slug_thuong_hieu'   =>  $request->slug_thuong_hieu,
-            'hinh_anh'          =>  $request->hinh_anh,
-            'tinh_trang'        =>  $request->tinh_trang,
-        ]);
 
-        return response()->json([
-            'status'    => true,
-            'message'   => 'Cập nhật mới thành công',
-        ]);
+        $who = Auth::guard('admin')->user();
+
+        $check = PermisionDetail::join('list_permisions', 'list_permisions.id', 'permision_details.id_hanh_dong')
+        ->where('id_quyen', $who->id_phan_quyen)
+            ->select('list_permisions.ten_hanh_dong')
+            ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
+            ->toArray(); // Chuyển đổi sang mảng
+        if (in_array('Quản Lý Thương Hiệu', $check)) {
+            $brand = Brand::where('id', $request->id)->first();
+            $brand->update([
+                'ten_thuong_hieu'    =>  $request->ten_thuong_hieu,
+                'slug_thuong_hieu'   =>  $request->slug_thuong_hieu,
+                'hinh_anh'          =>  $request->hinh_anh,
+                'tinh_trang'        =>  $request->tinh_trang,
+            ]);
+
+            return response()->json([
+                'status'    => true,
+                'message'   => 'Cập nhật mới thành công',
+            ]);
+        } else {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không đủ thẩm quyền để thực hiện chức năng này',
+            ]);
+        }
     }
     public function status(Request $request)
     {
-        $brand = Brand::where('id', $request->id)->first();
-        if ($brand) {
-            $brand->tinh_trang = !$brand->tinh_trang;
-            $brand->save();
+
+        $who = Auth::guard('admin')->user();
+
+        $check = PermisionDetail::join('list_permisions', 'list_permisions.id', 'permision_details.id_hanh_dong')
+        ->where('id_quyen', $who->id_phan_quyen)
+            ->select('list_permisions.ten_hanh_dong')
+            ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
+            ->toArray(); // Chuyển đổi sang mảng
+        if (in_array('Quản Lý Thương Hiệu', $check)) {
+            $brand = Brand::where('id', $request->id)->first();
+            if ($brand) {
+                $brand->tinh_trang = !$brand->tinh_trang;
+                $brand->save();
+            }
+            return response()->json([
+                'status'    => true,
+                'message'   => 'Đổi trạng thái thành công !',
+            ]);
+        } else {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không đủ thẩm quyền để thực hiện chức năng này',
+            ]);
         }
-        return response()->json([
-            'status'    => true,
-            'message'   => 'Đổi trạng thái thành công !',
-        ]);
     }
 }
