+++ app/Http/Controllers/ClassificationController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\createClassificationRequest;
 use App\Models\Classification;
+use App\Models\PermisionDetail;
 use App\Models\Vehicle;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ClassificationController extends Controller
 {
@@ -22,18 +24,32 @@ class ClassificationController extends Controller
     }
     public function store(createClassificationRequest $request)
     {
-        $classification = Classification::create($request->all());
-        if ($classification) {
-            return response()->json([
-                'status'    => 1,
-                'message'   => 'Thêm mới chỗ ngồi thành công!',
-            ]);
+        $who = Auth::guard('admin')->user();
+        $check = PermisionDetail::join('list_permisions', 'list_permisions.id', 'permision_details.id_hanh_dong')
+        ->where('id_quyen', $who->id_phan_quyen)
+            ->select('list_permisions.ten_hanh_dong')
+            ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
+            ->toArray(); // Chuyển đổi sang mảng
+        if (in_array('Quản Lý Xe', $check)) {
+            $classification = Classification::create($request->all());
+            if ($classification) {
+                return response()->json([
+                    'status'    => 1,
+                    'message'   => 'Thêm mới chỗ ngồi thành công!',
+                ]);
+            } else {
+                return response()->json([
+                    'status'    => 0,
+                    'message'   => 'Thêm mới chỗ ngồi thất bại!',
+                ]);
+            }
         } else {
             return response()->json([
                 'status'    => 0,
-                'message'   => 'Thêm mới chỗ ngồi thất bại!',
+                'message'   => 'Bạn không đủ thẩm quyền để thực hiện chức năng này',
             ]);
         }
+
     }
     public function search(Request $request)
     {
@@ -48,27 +64,42 @@ class ClassificationController extends Controller
     }
     public function delete(Request $request)
     {
-        $classification = Classification::find($request->id);
-        $vehicle = Vehicle::where('id_loai_xe', $request->id)->first();
+        $who = Auth::guard('admin')->user();
 
-        if ($vehicle) {
-            return response()->json([
-                'status'    => 0,
-                'message'   => 'Loại xe này đang được sử dụng, tạm thời không thể xoá!',
-            ]);
-        } else {
-            if ($classification) {
-                $classification->delete();
-                return response()->json([
-                    'status'    => 1,
-                    'message'   => 'Đã xoá thành công!',
-                ]);
-            } else {
+        $check = PermisionDetail::join('list_permisions', 'list_permisions.id', 'permision_details.id_hanh_dong')
+        ->where('id_quyen', $who->id_phan_quyen)
+            ->select('list_permisions.ten_hanh_dong')
+            ->pluck('ten_hanh_dong') // Lấy mảng của các giá trị 'ten_hanh_dong'
+            ->toArray(); // Chuyển đổi sang mảng
+        if (in_array('Quản Lý Xe', $check)) {
+            $classification = Classification::find($request->id);
+            $vehicle = Vehicle::where('id_loai_xe', $request->id)->first();
+
+            if ($vehicle) {
                 return response()->json([
                     'status'    => 0,
-                    'message'   => 'Loại xe này không tồn tại!',
+                    'message'   => 'Loại xe này đang được sử dụng, tạm thời không thể xoá!',
                 ]);
+            } else {
+                if ($classification) {
+                    $classification->delete();
+                    return response()->json([
+                        'status'    => 1,
+                        'message'   => 'Đã xoá thành công!',
+                    ]);
+                } else {
+                    return response()->json([
+                        'status'    => 0,
+                        'message'   => 'Loại xe này không tồn tại!',
+                    ]);
+                }
             }
+        } else {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không đủ thẩm quyền để thực hiện chức năng này',
+            ]);
         }
+
     }
 }
